photo null check

+++ app/Http/Controllers/Api/Profile/PhotoController.php
@@ -151,6 +151,9 @@ class PhotoController extends Controller
     public function destroy(Request $request, $profileId, $id)
     {
         $this->model =  $request->user()->profile->photos()->where('id',$id)->first();
+        if(!$this->model){
+            return $this->sendError("Photo not found.");
+        }
         event(new DeleteFeedable($this->model));
         $this->model = $this->model->delete();
         //remove from recent photos
